requirements.txt
flask
ikea-api
pandas
pillow
pydantic
requests
torch
transformers
